Add two segment images of central flooded roads
Used to answer sub-question c).

Two notable combinations of two high-centrality edges that eventually will be flooded are the (1543, 1777)-(1777, 314)-combination and the (956, 957)-(957, 1256)-combination. The former are two segments of the type of primary_link, meaning highway entrances, exits or transfers. In this case, it’s a highway entrance and exit that are used to avoid a flood on the (1543, 314)-segment of the Hakamäentie, the Finnish regional road 100. Such a turn wouldn’t be possible, so this is a model limitation.
The second combination is a secondary road that is used significantly starting from the 44 mm scenario, and it the most used road in the 54 and 71 mm scenarios. The secondary road becomes the fastest route after segments on the Tuusulanväylä and Kehä III highways flood. It stops being used in the 86 mm scenario, when both segments are flooded.